Start of uma2 module, dividing up the UMA2 mega file
