+++ src/provider.rs
@@ -26,24 +26,6 @@ pub trait Provider {
     }
 }
 
-#[cfg(feature = "uma2")]
-pub trait Uma2Provider {
-    /// Whether UMA2 capabilities have been discovered
-    fn uma2_discovered(&self) -> bool;
-
-    /// UMA-compliant Resource Registration Endpoint which resource servers can use to manage their
-    /// protected resources and scopes. This endpoint provides operations create, read, update and
-    /// delete resources and scopes
-    fn resource_registration_uri(&self) -> Option<&Url>;
-
-    /// UMA-compliant Permission Endpoint which resource servers can use to manage permission
-    /// tickets. This endpoint provides operations create, read, update, and delete permission tickets
-    fn permission_uri(&self) -> Option<&Url>;
-
-    /// API from where permissions can be set to resources by resource servers on behalf of their users.
-    fn uma_policy_uri(&self) -> Option<&Url>;
-}
-
 /// Google OAuth 2.0 providers.
 ///
 /// See [Using OAuth 2.0 to Access Google
